getauxval: Try glibc's `getauxval` before trying `/proc/self/auxv`
First, try loading and using `getauxval` from glibc by using
`dlsym(RTLD_DEFAULT, "getauxval")`. `getauxval` was added in glibc 2.16
(2012-06-30); it had a bug in which it failed to report `ENOENT` for
missing elf auxiliary vector "types". This bug was fixed in glibc 2.19
(2014-02-07).

Second, try reading auxiliary vector values from `/proc/self/auxv`. The
contents of this file are copied to a local 64-entry buffer; it is
extremely unlikely that a system using glibc < 2.16 will also use a
kernel that has more than 64 auxv entries. It is possible for procfs to
not be mounted, or for some procfs entries to be unreadable (like on
e.g. Android [1], although Android is not a target platform for this
project).

Finally, if the above doesn't work, fall back to a function that always
returns 0 and sets `ENOENT`.

glibc exports a dynamic symbol named `_argv` that points to the original
argument vector. It would be possible to fetch the auxiliary vector by
getting the address of `_argv` via `dlsym(RTLD_DEFAULT, "_argv")` and
then scanning forward for the argv zero terminator and environment zero
terminator.  Unfortunately, `_argc` is not exported, and the memory
pointed to by `_argv` may be mutated arbitrarily; it's possible to read
`/proc/self/cmdline` to determine the size of the argument vector (and
`/proc/self/environ` to determine the size of the memory pointed to by
`__environ`), but if procfs is accessible and glibc is being used (e.g.
if the underlying os is normal Linux and not Android) then
`/proc/self/auxv` should also be readable.

It'd also be possible to add before-main startup code that saves a copy
of the address of the auxiliary vector that is then read by an exported
(default visibility) getauxval-compatible function; in this case, an
executable using musl-glibc, and dlls built with musl-glibc loaded in a
musl-glibc executable, would have a reliable implementation of
`getauxval`. I may do this later.

[1] https://android-review.googlesource.com/c/platform/ndk/+/51271/1/sources/android/cpufeatures/cpu-features.c